active-projects/MonthlyRatioAutomation/requirements.txt
requests==2.24.0
numpy==1.14.6
beautifulsoup4==4.9.3
camelot==13.04.13-gpl-pyqt
envparse==0.2.0
pandas==1.2.4
pyjarowinkler==1.8
python_dateutil==2.8.1
SQLAlchemy==1.4.15
